heroku latest version

+++ server/routes/product.js
@@ -1,41 +1,48 @@
 const express = require('express');
 const router = express.Router();
-const multer = require('multer');
 const { Product } = require('../models/Product');
-const fs = require('fs');
+
+const multer = require('multer');
+const multerS3 = require('multer-s3')
+const AWS = require('aws-sdk');
+const awsConfig = require("../config/awsConfig");
 
 //=================================
 //             Product
 //=================================
 
-// 이미지파일 등록 라이브러리 정의
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, 'uploads/')
-  },
-  filename: function (req, file, cb) {
-    //const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9)
-    cb(null, `${Date.now()}_${file.originalname}`); //file.fieldname + '-' + uniqueSuffix)
-  }
-})
+// AWS S3 접속 Key
+const s3 = new AWS.S3({
+  accessKeyId: awsConfig.accessKeyId,
+  secretAccessKey: awsConfig.secretAccessKey,
+  region: awsConfig.region
+});
 
-const upload = multer({ storage: storage }).single("file");
+// 이미지파일 등록 AWS 정의
+const upload = multer({
+    storage: multerS3({
+      s3: s3,
+      bucket: "furuokadrug-bucket",
+      metadata: function (req, file, cb) {
+        cb(null, { fieldName: file.fieldname});
+      },
+      key: function (req, file, cb) {
+        cb(null, `${ Date.now()}_${file.originalname }`);
+      },
+    }),
+  });
 
-// 이미지파일 등록
+// 이미지 등록
 router.post('/image', (req, res) => {
-  // 가져온 이미지를 저장한다
-  upload(req, res, err => {
-    if(err) {
-      return res.json({success: false, err});
-    }
-    return res.json({ success: true, filePath: res.req.file.path, fileName: res.req.file.filename});
+  const uploadSingle = upload.single("file");
+  uploadSingle(req, res, err => {
+    if(err) return res.json({success: false, err});
+    return res.json({ success: true, filePath: res.req.file.location});
   })
 })
 
-
-// 상품 등록(이미지 포함)
+// 상품 등록(이미지 경로포함)
 router.post('/', (req, res) => {
-  // 받아온 정보들을 디비에 넣어준다
   const product = new Product(req.body)
   product.save((err) => {
     if (err) return res.status(400).json({success: false, err})
@@ -43,17 +50,18 @@ router.post('/', (req, res) => {
   })
 })
 
-// 상품 가져오기
+// 상품 가져오기(LandingPage)
 router.post('/products', (req, res) => {
-  // product collection에 들어 있는 모든 상품 정보 가져오기
+  // 화면 표시 갯수 정의
   let limit = req.body.limit ? parseInt(req.body.limit) : 20;
-  let skip  = req.body.skip  ? parseInt(req.body.skip)  : 0;
+  let skip  = req.body.skip ? parseInt(req.body.skip)  : 0;
+
   let term = req.body.searchTerm;
+  // findArgs: find에 객체 형태로 만들어 Query의 조건식으로 사용
   let findArgs = {};
   
   for (let key in req.body.filters) {
     if (req.body.filters[key].length > 0) {
-
       // key -> continent 또는 price
       if (key === "price") {
         findArgs[key] = {
@@ -82,7 +90,6 @@ router.post('/products', (req, res) => {
         return res.status(200).json({ success: true, productInfo, postSize: productInfo.length })
       })
   } else {
-    // findArgs: find에 객체 형태로 주면 조건식이 된다
     Product.find(findArgs)
       .populate("writer")
       .skip(skip)
@@ -97,7 +104,7 @@ router.post('/products', (req, res) => {
 // 상품상세(상품정보 가져오기), 카트페이지에서 상폼정보 가져오기
 router.get('/products_by_id', (req, res) => {
   let type = req.query.type;
-  let productIds = req.query.id;
+  let productId = req.query.id;
 
   if (type === "array") {
     let ids = req.query.id.split(',')
@@ -113,42 +120,73 @@ router.get('/products_by_id', (req, res) => {
 
 // 상품삭제
 router.post('/delete', (req, res) => {  
-  // Image 삭제
+  let isErr = false;
+
   if (req.body.images.length > 0) {
     let images = req.body.images;
+    
+    console.log("images: ", images);
+
+    // AWS S3 이미지 삭제
     for (let i=0; i<images.length; i++) {
-      let path = images[i];
-      try {
-        fs.unlinkSync(path);
-      } catch(err) {
-        return res.json({ success: false}, err);
-      }
+      let str = images[i];
+      let words = str.split('/');
+      let fileName = words[words.length-1]
+
+      console.log("fileName: ", fileName);
+
+      s3.deleteObject({
+        Bucket: "furuokadrug-bucket", // 사용자 버켓 이름
+        Key: fileName // 버켓 내 경로
+      }, (err, data) => {
+        if (err) { 
+          isErr = true;
+          console.log(err, err.stack);
+        }
+        console.log("image delete: ", data);
+      })
     } 
   }
 
-  if (req.body.id) {
+  if (!isErr) {
     // DataBase 삭제
-    Product.deleteOne({_id: req.body.id})
-    .then((r)=>{
-      console.log(r.result);
-      return res.json({ success: true});
-    }, (err)=>{
-      if(err) console.log(err);
-      return res.json({success: false, err});
-    });
+    if (req.body.id) {
+      Product.deleteOne({_id: req.body.id})
+      .then((r)=>{
+        console.log("DB delete: ", r.result);
+      }, (err)=>{
+        if(err) {
+          console.log(err);
+          isErr = true;
+        }
+      });
+    }
+  }
+
+  console.log("isErr3: ", isErr);
+  if (!isErr) {
+    return res.json({success: true});
+  } else {
+    return res.json({success: false});
   }
 })
 
-// Image 삭제
+// 상품 등록에서 이미지삭제
 router.post('/delete_image', (req, res) => {
-  let path = req.body.image;
-  try {
-    fs.unlinkSync(path);
-    return res.json({ success: true});
-  } catch(err) {
-    return res.json({ success: false}, err);
-  }
-})
+  let str = req.body.image;
+  let words = str.split('/');
+  let fileName = words[words.length-1]
 
+  s3.deleteObject({
+    Bucket: "furuokadrug-bucket", // 사용자 버켓 이름
+    Key: fileName // 버켓 내 경로
+  }, (err, data) => {
+    if (err) { 
+      console.log(err, err.stack);
+      return res.json({ success: false }) 
+    }
+    return res.json({ success: true});    
+  })
+})
 
 module.exports = router;
